fix: `print`s are not working from inside a resource in simulator
Fixes #1871 by rewriting our implementation for `require()` function inside the inflight sandbox. The previous implementation allowed modules that were required from within the VM to escape the VM's context. This resulted in the `console.log` hijack to not carry over to other modules (such as inflight client modules of resources).

The new implementation resolves the required module relative to the parent module, and evaluates its contents inside the VM with the same sandbox context.

This means that now we have more control over how to resolve modules within the VM, so technically we should be able to add custom module resolution paths and relax the constraint that required the simulator output to reside below the project's `node_modules` directory, but that's not required right now.